Changed load_image() to use matlab's memoize function capabilities
The input table look up for the memoized function is terrible.  It
does an unordered search through the table, so trying to load N
images from a cache holding N images takes N^2 time, which comes
out to be roughly 48 seconds for N=6000.  For comparison, not
using a cache and just reading the data from the disk takes about
15 seconds.  Furthermore, just returning the data from the output
cache takes around 15 seconds, so it's not great.

To do this properly I'd have to write my own memoization code that
uses a hash table for faster look ups.  Probably can use matlab's
map class.